replaced activity navigation with jetpack fragment navigation
