+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     package="agersant.polaris">
 
@@ -10,43 +9,38 @@
     <application
         android:name=".PolarisApplication"
         android:allowBackup="true"
-        android:fullBackupOnly="false"
         android:fullBackupContent="@xml/backup_rules"
+        android:fullBackupOnly="false"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/Polaris"
         android:usesCleartextTraffic="true"
-        tools:ignore="GoogleAppIndexingWarning"
-        >
+        tools:ignore="GoogleAppIndexingWarning">
 
-        <activity android:name=".features.browse.CollectionActivity">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".features.settings.SettingsActivity" />
-        <activity android:name=".features.browse.BrowseActivity" />
-        <activity android:name=".features.queue.QueueActivity" />
-        <activity android:name=".features.player.PlayerActivity" android:parentActivityName=".features.browse.CollectionActivity" />
 
         <service
             android:name=".PolarisPlaybackService"
-            android:enabled="true"
             android:description="@string/media_player_service_description"
+            android:enabled="true"
             android:exported="false" />
 
         <service
             android:name=".PolarisDownloadService"
-            android:enabled="true"
             android:description="@string/download_service_description"
+            android:enabled="true"
             android:exported="false" />
 
         <service
             android:name=".PolarisScrobbleService"
-            android:enabled="true"
             android:description="@string/scrobble_service_description"
+            android:enabled="true"
             android:exported="false" />
 
     </application>
