--- app/src/main/res/layout-land/activity_player.xml
@@ -1,122 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_player"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context="agersant.polaris.features.player.PlayerActivity">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <com.google.android.material.appbar.MaterialToolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize" />
-
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:orientation="horizontal">
-
-        <agersant.polaris.ui.SquareLayout
-            android:id="@+id/viewGroup"
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            app:preserveWidth="false">
-
-            <ImageView
-                android:id="@+id/artwork"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:adjustViewBounds="true"
-                android:contentDescription="@string/player_artwork"
-                tools:src="@drawable/artwork" />
-        </agersant.polaris.ui.SquareLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:gravity="center_vertical"
-            android:orientation="vertical">
-
-            <LinearLayout
-                android:id="@+id/controls"
-                android:layout_width="match_parent"
-                android:layout_height="48dp"
-                android:gravity="center_horizontal"
-                android:orientation="horizontal">
-
-                <ImageView
-                    android:id="@+id/skip_previous"
-                    android:layout_width="wrap_content"
-                    android:layout_height="match_parent"
-                    android:adjustViewBounds="true"
-                    android:contentDescription="@string/player_previous_track"
-                    android:onClick="skipPrevious"
-                    android:src="@drawable/ic_skip_previous_black_24dp"
-                    android:tint="?android:textColorPrimary" />
-
-                <ImageView
-                    android:id="@+id/pause_toggle"
-                    android:layout_width="wrap_content"
-                    android:layout_height="match_parent"
-                    android:layout_marginLeft="32dp"
-                    android:layout_marginRight="32dp"
-                    android:adjustViewBounds="true"
-                    android:contentDescription="@string/player_play_pause"
-                    android:onClick="togglePause"
-                    android:src="@drawable/ic_play_arrow_black_24dp"
-                    android:tint="?android:textColorPrimary" />
-
-                <ImageView
-                    android:id="@+id/skip_next"
-                    android:layout_width="wrap_content"
-                    android:layout_height="match_parent"
-                    android:adjustViewBounds="true"
-                    android:contentDescription="@string/player_next_track"
-                    android:onClick="skipNext"
-                    android:src="@drawable/ic_skip_next_black_24dp"
-                    android:tint="?android:textColorPrimary" />
-            </LinearLayout>
-
-            <TextView
-                android:id="@+id/buffering"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="16dp"
-                android:layout_marginTop="-12dp"
-                android:layout_marginEnd="16dp"
-                android:layout_marginBottom="-3dp"
-                android:text="@string/player_buffering"
-                android:textAlignment="textStart"
-                android:textSize="12sp" />
-
-            <SeekBar
-                android:id="@+id/seek_bar"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:max="100"
-                tools:progress="20"
-                tools:secondaryProgress="20" />
-
-        </LinearLayout>
-    </LinearLayout>
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/navigation"
-        android:layout_width="match_parent"
-        android:layout_height="56dp"
-        android:background="?colorHeaderBackground"
-        app:itemBackground="?colorHeaderBackground"
-        app:itemIconTint="@drawable/bottom_navigation_item"
-        app:itemTextColor="@drawable/bottom_navigation_item"
-        app:menu="@menu/menu_navigation" />
-
-</LinearLayout>
