+++ app/src/main/res/layout-land/fragment_player.xml
@@ -1,0 +1,89 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="horizontal">
+
+    <agersant.polaris.ui.SquareLayout
+        android:id="@+id/viewGroup"
+        android:layout_width="0dp"
+        android:layout_height="match_parent"
+        app:preserveWidth="false">
+
+        <ImageView
+            android:id="@+id/artwork"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:adjustViewBounds="true"
+            android:contentDescription="@string/player_artwork"
+            tools:src="@drawable/artwork" />
+    </agersant.polaris.ui.SquareLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:gravity="center_vertical"
+        android:orientation="vertical">
+
+        <LinearLayout
+            android:id="@+id/controls"
+            android:layout_width="match_parent"
+            android:layout_height="48dp"
+            android:gravity="center_horizontal"
+            android:orientation="horizontal">
+
+            <ImageView
+                android:id="@+id/skip_previous"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:adjustViewBounds="true"
+                android:contentDescription="@string/player_previous_track"
+                android:src="@drawable/ic_skip_previous_black_24dp"
+                android:tint="?colorOnBackground" />
+
+            <ImageView
+                android:id="@+id/pause_toggle"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_marginLeft="32dp"
+                android:layout_marginRight="32dp"
+                android:adjustViewBounds="true"
+                android:contentDescription="@string/player_play_pause"
+                android:src="@drawable/ic_play_arrow_black_24dp"
+                android:tint="?colorOnBackground" />
+
+            <ImageView
+                android:id="@+id/skip_next"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:adjustViewBounds="true"
+                android:contentDescription="@string/player_next_track"
+                android:src="@drawable/ic_skip_next_black_24dp"
+                android:tint="?colorOnBackground" />
+        </LinearLayout>
+
+        <TextView
+            android:id="@+id/buffering"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="16dp"
+            android:layout_marginTop="-12dp"
+            android:layout_marginEnd="16dp"
+            android:layout_marginBottom="-3dp"
+            android:text="@string/player_buffering"
+            android:textAlignment="textStart"
+            android:textSize="12sp" />
+
+        <SeekBar
+            android:id="@+id/seek_bar"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:max="100"
+            tools:progress="20"
+            tools:secondaryProgress="20" />
+
+    </LinearLayout>
+
+</LinearLayout>
