--- app/src/main/res/layout/activity_browse.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                xmlns:app="http://schemas.android.com/apk/res-auto"
-                xmlns:tools="http://schemas.android.com/tools"
-                android:id="@+id/activity_browse"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                tools:context="agersant.polaris.features.browse.BrowseActivity">
-
-	<com.google.android.material.appbar.AppBarLayout
-		android:id="@+id/appbar_layout"
-		android:layout_width="match_parent"
-		android:layout_height="wrap_content">
-
-		<com.google.android.material.appbar.MaterialToolbar
-			android:id="@+id/toolbar"
-			android:layout_width="match_parent"
-			android:layout_height="?attr/actionBarSize" />
-
-	</com.google.android.material.appbar.AppBarLayout>
-
-	<ProgressBar
-		android:id="@+id/progress_bar"
-		android:layout_width="match_parent"
-		android:layout_height="48dp"
-		android:layout_centerVertical="true"
-		android:indeterminateTint="?colorPrimary"
-		tools:visibility="gone"/>
-
-	<LinearLayout
-		android:id="@+id/browse_error_message"
-		android:layout_width="match_parent"
-		android:layout_height="wrap_content"
-		android:layout_centerVertical="true"
-		android:layout_centerHorizontal="true"
-		android:gravity="center"
-		android:layout_marginStart="8dp"
-		android:layout_marginEnd="8dp"
-		android:orientation="vertical">
-		<TextView
-			android:layout_width="wrap_content"
-			android:layout_height="wrap_content"
-			android:gravity="center"
-			android:textSize="16sp"
-			android:text="@string/browse_error"
-			android:drawableStart="@drawable/ic_error_black_24dp"
-			android:drawableTint="?colorError"
-			android:drawablePadding="8dp"/>
-		<Button
-			android:id="@+id/browse_error_retry"
-			android:layout_width="80dp"
-			android:layout_height="40dp"
-			android:textColor="?android:textColorPrimary"
-			android:onClick="retry"
-			android:text="@string/browse_error_retry"/>
-	</LinearLayout>
-
-	<FrameLayout
-		android:id="@+id/browse_content_holder"
-		android:layout_width="match_parent"
-		android:layout_height="match_parent"
-		android:layout_above="@+id/navigation"
-		android:layout_below="@+id/appbar_layout"/>
-
-	<com.google.android.material.bottomnavigation.BottomNavigationView
-		android:id="@+id/navigation"
-		android:layout_width="match_parent"
-		android:layout_height="56dp"
-		android:layout_alignParentBottom="true"
-		android:background="?colorHeaderBackground"
-		app:itemBackground="?colorHeaderBackground"
-		app:itemIconTint="@drawable/bottom_navigation_item"
-		app:itemTextColor="@drawable/bottom_navigation_item"
-		app:menu="@menu/menu_navigation"/>
-
-</RelativeLayout>
