--- app/src/main/res/layout/activity_collection.xml
@@ -1,84 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="agersant.polaris.features.browse.CollectionActivity">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:id="@+id/appbar_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <com.google.android.material.appbar.MaterialToolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize" />
-
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@id/appbar_layout"
-        android:orientation="vertical"
-        android:paddingLeft="8dp"
-        android:paddingRight="8dp">
-
-        <Button
-            android:id="@+id/browse_directories"
-            style="@style/CollectionButton"
-            android:layout_width="match_parent"
-            android:layout_height="48dp"
-            android:drawableStart="@drawable/ic_folder_black_24dp"
-            android:onClick="browseDirectories"
-            android:text="@string/collection_browse_directories" />
-
-        <View style="@style/Divider" />
-
-        <Button
-            android:id="@+id/random"
-            style="@style/CollectionButton"
-            android:layout_width="match_parent"
-            android:layout_height="48dp"
-            android:drawableStart="@drawable/ic_shuffle_black_24dp"
-            android:onClick="browseRandom"
-            android:text="@string/collection_random_albums" />
-
-        <View style="@style/Divider" />
-
-        <Button
-            android:id="@+id/recently_added"
-            style="@style/CollectionButton"
-            android:layout_width="match_parent"
-            android:layout_height="48dp"
-            android:drawableStart="@drawable/ic_new_releases_black_24dp"
-            android:onClick="browseRecent"
-            android:text="@string/collection_recently_added" />
-
-        <View style="@style/Divider" />
-
-        <Button
-            android:id="@+id/playlists"
-            style="@style/CollectionButton"
-            android:layout_width="match_parent"
-            android:layout_height="48dp"
-            android:drawableStart="@drawable/ic_queue_music_black_24dp"
-            android:text="@string/collection_playlists" />
-
-    </LinearLayout>
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/navigation"
-        android:layout_width="match_parent"
-        android:layout_height="56dp"
-        android:layout_alignParentBottom="true"
-        android:background="?colorHeaderBackground"
-        app:itemBackground="?colorHeaderBackground"
-        app:itemIconTint="@drawable/bottom_navigation_item"
-        app:itemTextColor="@drawable/bottom_navigation_item"
-        app:menu="@menu/menu_navigation" />
-
-</RelativeLayout>
