+++ app/src/main/res/layout/fragment_browse.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/activity_browse"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="agersant.polaris.features.browse.BrowseFragment">
+
+    <ProgressBar
+        android:id="@+id/progress_bar"
+        android:layout_width="match_parent"
+        android:layout_height="48dp"
+        android:layout_centerVertical="true"
+        android:indeterminateTint="?colorPrimary"
+        tools:visibility="gone" />
+
+    <LinearLayout
+        android:id="@+id/browse_error_message"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_centerHorizontal="true"
+        android:layout_centerVertical="true"
+        android:layout_marginStart="8dp"
+        android:layout_marginEnd="8dp"
+        android:gravity="center"
+        android:orientation="vertical">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:drawablePadding="8dp"
+            android:gravity="center"
+            android:text="@string/browse_error"
+            android:textSize="16sp"
+            app:drawableStartCompat="@drawable/ic_error_black_24dp"
+            app:drawableTint="?colorError" />
+
+        <Button
+            android:id="@+id/browse_error_retry"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:minWidth="0dp"
+            android:minHeight="0dp"
+            android:text="@string/browse_error_retry"
+            android:textColor="?android:textColorPrimary" />
+    </LinearLayout>
+
+    <FrameLayout
+        android:id="@+id/browse_content_holder"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
+
+</RelativeLayout>
