+++ app/src/main/res/layout/fragment_queue.xml
@@ -1,30 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/activity_queue"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="agersant.polaris.features.queue.QueueActivity">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:id="@+id/appbar_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <com.google.android.material.appbar.MaterialToolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize" />
-
-    </com.google.android.material.appbar.AppBarLayout>
+    tools:context="agersant.polaris.features.queue.QueueFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/queue_recycler_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_above="@+id/navigation"
-        android:layout_below="@id/appbar_layout"
         android:scrollbars="vertical"
         tools:listitem="@layout/view_queue_item" />
 
@@ -58,16 +43,5 @@
             android:textSize="16sp" />
 
     </LinearLayout>
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/navigation"
-        android:layout_width="match_parent"
-        android:layout_height="56dp"
-        android:layout_alignParentBottom="true"
-        android:background="?colorHeaderBackground"
-        app:itemBackground="?colorHeaderBackground"
-        app:itemIconTint="@drawable/bottom_navigation_item"
-        app:itemTextColor="@drawable/bottom_navigation_item"
-        app:menu="@menu/menu_navigation" />
 
 </RelativeLayout>
