+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/nav_collection">
+
+    <fragment
+        android:id="@+id/nav_settings"
+        android:name="agersant.polaris.features.settings.SettingsFragment"
+        android:label="@string/settings" />
+
+    <fragment
+        android:id="@+id/nav_collection"
+        android:name="agersant.polaris.features.browse.CollectionFragment"
+        android:label="@string/collection"
+        tools:layout="@layout/fragment_collection">
+
+        <action
+            android:id="@+id/action_nav_collection_to_nav_browse"
+            app:destination="@id/nav_browse" />
+
+    </fragment>
+
+    <fragment
+        android:id="@+id/nav_browse"
+        android:name="agersant.polaris.features.browse.BrowseFragment"
+        android:label="@string/collection"
+        tools:layout="@layout/fragment_browse" />
+
+    <fragment
+        android:id="@+id/nav_queue"
+        android:name="agersant.polaris.features.queue.QueueFragment"
+        android:label="@string/queue"
+        tools:layout="@layout/fragment_queue" />
+
+    <fragment
+        android:id="@+id/nav_now_playing"
+        android:name="agersant.polaris.features.player.PlayerFragment"
+        android:label="@string/now_playing"
+        tools:layout="@layout/fragment_player" />
+
+</navigation>
